third_party/WebKit/LayoutTests/http/tests/payments/payment-app-interfaces.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
<script>
  // Tests that the expected Payment App API interfaces are being exposed on both
  // the Service Worker and Document global scopes.
  service_worker_test(
      'resources/payment-app.js',
      'Exposure of interfaces in a Service Worker.');

  test(function() {
      assert_own_property(self, 'PaymentManager', 'PaymentManager needs to be exposed as a global.');
  }, 'PaymentManager should be exposed and have the expected interface in a Document.');
</script>
